Move packets into separate file
