+++ src/multiplayer_connection.h
@@ -9,7 +9,10 @@
 #include <functional>
 #include <type_traits>
 #include <string>
-#include <charconv>
+
+#include "multiplayer_packet.h"
+
+namespace Multiplayer {
 
 class MessageProcessingException : public std::runtime_error {
 public:
@@ -18,89 +21,15 @@ public:
 	MessageProcessingException(const std::string& w) : runtime_error(std::move(w)) {}
 };
 
-class MultiplayerConnection {
+class Connection {
 public:
-	constexpr static std::string_view PARAM_DELIM = "\uFFFF";
-	constexpr static std::string_view MSG_DELIM = "\uFFFE";
-
-	MultiplayerConnection() = default;
-	MultiplayerConnection(const MultiplayerConnection&) = delete;
-	MultiplayerConnection(MultiplayerConnection&&) = default;
-	MultiplayerConnection& operator=(const MultiplayerConnection&) = delete;
-	MultiplayerConnection& operator=(MultiplayerConnection&&) = default;
+	Connection() = default;
+	Connection(const Connection&) = delete;
+	Connection(Connection&&) = default;
+	Connection& operator=(const Connection&) = delete;
+	Connection& operator=(Connection&&) = default;
 
 	using ParameterList = std::vector<std::string_view>;
-
-	class Packet {
-	public:
-		Packet() {}
-		virtual ~Packet() = default;
-	protected:
-	};
-
-	class C2SPacket : public Packet {
-	public:
-		virtual ~C2SPacket() = default;
-		virtual std::string ToBytes() const = 0;
-
-		C2SPacket(std::string _name) : m_name(std::move(_name)) {}
-		std::string_view GetName() const { return m_name; }
-
-		static std::string Sanitize(std::string_view param);
-
-		static std::string ToString(const char* x) { return ToString(std::string_view(x)); }
-		static std::string ToString(int x) { return std::to_string(x); }
-		static std::string ToString(bool x) { return x ? "1" : "0"; }
-		static std::string ToString(std::string_view v) { return Sanitize(v); }
-
-		template<typename... Args>
-		std::string Build(Args... args) const {
-			std::string prev {m_name};
-			AppendPartial(prev, args...);
-			return prev;
-		}
-
-		template<typename T>
-		static void AppendPartial(std::string& s, T t) {
-			s += PARAM_DELIM;
-			s += ToString(t);
-		}
-
-		template<typename T, typename... Args>
-		static void AppendPartial(std::string& s, T t, Args... args) {
-			s += PARAM_DELIM;
-			s += ToString(t);
-			AppendPartial(s, args...);
-		}
-	protected:
-		std::string m_name;
-	};
-
-	class S2CPacket : public Packet {
-	public:
-		virtual ~S2CPacket() = default;
-
-		template<typename T>
-		static T Decode(std::string_view s);
-
-		template<>
-		int Decode(std::string_view s) {
-			int r;
-			auto e = std::from_chars(s.data(), s.data() + s.size(), r);
-			if (e.ec != std::errc())
-				throw MessageProcessingException("Decoding int");
-			return r;
-		}
-
-		template<>
-		bool Decode(std::string_view s) {
-			if (s == "1")
-				return true;
-			if (s == "0")
-				return true;
-			throw MessageProcessingException("Decoding bool");
-		}
-	};
 
 	void SendPacket(const C2SPacket& p);
 	template<typename T, typename... Args>
@@ -130,19 +59,19 @@ public:
 		CLOSE,
 		_PLACEHOLDER,
 	};
-	using SystemMessageHandler = std::function<void (MultiplayerConnection&)>;
+	using SystemMessageHandler = std::function<void (Connection&)>;
 	void RegisterSystemHandler(SystemMessage m, SystemMessageHandler h);
 
 	void Dispatch(std::string_view name, ParameterList args = ParameterList());
 
 	bool IsConnected() const { return connected; }
 
-	virtual ~MultiplayerConnection() = default;
+	virtual ~Connection() = default;
 
 	void SetKey(std::string k) { key = std::move(k); }
 	std::string_view GetKey() const { return key; }
 
-	static std::vector<std::string_view> Split(std::string_view src, std::string_view delim = PARAM_DELIM);
+	static std::vector<std::string_view> Split(std::string_view src, std::string_view delim = Packet::PARAM_DELIM);
 
 protected:
 	bool connected;
@@ -156,6 +85,8 @@ protected:
 
 	std::string key;
 };
+
+}
 
 #endif
